+++ src/main/java/shopmod/screens/RelicPopUp.java
@@ -133,10 +133,11 @@ public class RelicPopUp {
             this.hbTop.render(sb);
             this.hbBot.render(sb);
             if (this.hbBot.hovered) {
+                float tipX = this.x > Settings.WIDTH * 0.75f ? this.x - (124.0f + 320.0f) * Settings.scale : this.x + 124.0f * Settings.scale;
                 if (MerchantsRug.canSell(relic)) {
-                    TipHelper.renderGenericTip(this.x + 124.0f * Settings.scale, this.y + 50.0f * Settings.scale, SELL_LABEL, SELL_MESSAGE[0] + MerchantsRug.relicSalePrice(relic) + SELL_MESSAGE[1]);
+                    TipHelper.renderGenericTip(tipX, this.y + 50.0f * Settings.scale, SELL_LABEL, SELL_MESSAGE[0] + MerchantsRug.relicSalePrice(relic) + SELL_MESSAGE[1]);
                 } else {
-                    TipHelper.renderGenericTip(this.x + 124.0f * Settings.scale, this.y + 50.0f * Settings.scale, CANT_SELL_LABEL, CANT_SELL_MESSAGE[0]);
+                    TipHelper.renderGenericTip(tipX, this.y + 50.0f * Settings.scale, CANT_SELL_LABEL, CANT_SELL_MESSAGE[0]);
                 }
             }
         }