+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v1.0.3
 * Fix: correctly position relic popup if there are multiple pages of relics.
+* Fix: relic sell price tip no longer moves off-screen for relics on the right of the screen.
+* Fix: when the relic popup is open, don't animate cards behind it.
 
 ### v1.0.2
 * Fix: selling relics no longer buys the card behind the click